Ports not unplugged following unbind
When using the networking genericswitch ML2 driver with ironic, neutron
ports on a tenant network are not unplugged after they have been
unbound. This means that the corresponding port remains an untagged
member of the tenant network VLAN on the switch.

This was seen on a Dell/Force10 S6010-ON switch. These switches have
slightly different CLI semantics than Cisco switches, which makes the
plug/unplug operation non-idempotent. The non-idempotent nature of the
operation means that a port that is an access port on one VLAN cannot be
made an access port on another VLAN using the 'plug' command alone - the
'unplug' command is required as well. This means that without manual
intervention, port configuration becomes a 'one shot' activity.

The NGS driver currently unplugs ports from VLANs only when ports are
deleted. However, since (at least) Ocata, ironic unbinds[0] ports before
they are deleted. This means that they lose their port binding
information during an update operation, and therefore do not have the
information required to unplug the port on the switch during the delete
operation.

The solution to this is to add support for unplugging ports following an
unbind operation. This change adds that support to the
update_port_postcommit method of the NGS mechanism driver.

[0]
https://github.com/openstack/ironic/blob/19ced862a782f204048ca2cb7f4904d75a6d7925/ironic/drivers/modules/network/neutron.py#L199

Change-Id: Ie4ea148793f9a754fae415c2661c6f2add6a97c8
Closes-Bug: #1683687